update authentication process
